Create README.md
An  hacked version of the PyPDF2 PDF handler amended for Python 3, with additional compression routines, extended character set handling (extract to text works in many cases where it did not before) and with additional features to support my own version of the wx.lib.pdfviewer.

However 

my code is not the best - for example my variable naming conventions change from minute to minute 
NOT designed to be python 3 compatible 
Not tested in a structured manner - but have fired a variety of real world PDFs at my PDF viewer and it mostly worked
 


Probably will NOT work with Python 2.  Not my best code.  Test coverage is